Big timer release change
The timer release is only setup when a job is finished.
It it changed to support job queue. In timer callback,
the heap is scanned to release new jobs and setup for
the next release. For this reason, the callback might
be still running when reinit_release_heap is called.
As long as the locks are taking care of, there should
not be any deadlock.

Also commented out some debugging info.